Move confirmationBody to ApproveOrderService
